# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HelpDesk API",
        "description": "API REST para gerenciamento de chamados de suporte técnico.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Listar chamados",
                "description": "Retorna os chamados de acordo com o perfil do usuário autenticado.",
                "operationId": "listTickets",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar chamados por status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "aberto",
                                "em_atendimento",
                                "aguardando",
                                "resolvido",
                                "fechado"
                            ]
                        },
                        "example": "aberto"
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtrar chamados por prioridade",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "baixa",
                                "media",
                                "alta"
                            ]
                        },
                        "example": "alta"
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrar chamados por categoria",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de chamados"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Criar chamado",
                "description": "Cria um novo chamado para o usuário autenticado.",
                "operationId": "createTicket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "title",
                                    "description",
                                    "priority"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Computador não liga"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "O computador não apresenta nenhum sinal."
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "media"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chamado criado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tickets/{ticket}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Buscar chamado",
                "description": "Retorna um chamado específico.",
                "operationId": "showTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID do chamado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chamado encontrado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Chamado não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Atualizar chamado",
                "description": "Atualiza os dados básicos de um chamado.",
                "operationId": "updateTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID do chamado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Computador não liga"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição atualizada."
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "alta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chamado atualizado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tickets/{ticket}/status": {
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Alterar status do chamado",
                "description": "Altera o status seguindo o fluxo definido do HelpDesk.",
                "operationId": "changeTicketStatus",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID do chamado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "resolvido"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status alterado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "422": {
                        "description": "Transição inválida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tickets/{ticket}/assign": {
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Atribuir chamado",
                "description": "Atribui um chamado a um usuário com perfil de atendente ou administrador.",
                "operationId": "assignTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID do chamado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assigned_to"
                                ],
                                "properties": {
                                    "assigned_to": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chamado atribuído"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Informe o token Bearer obtido no login.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}